+++ src/Components/Input/DurationInput.jsx
@@ -25,7 +25,7 @@ export default class DurationInput extends TokenInput {
   }
 
   onFocus = () => {
-    this.props.onFocus();
+    this.props.onFocus(this.props.tokenKey);
     window.document.addEventListener('keydown', this.handleKeyDown);
   };