+++ src/Components/Input/SelectInput.jsx
@@ -40,7 +40,7 @@ export default class SelectInput extends TokenInput {
   }
 
   onFocus = () => {
-    this.props.onFocus();
+    this.props.onFocus(this.props.tokenKey);
     if (this.props.showSearch && this.searchInput) {
       this.searchInput.focus();
     }